Main (#98)
* Feature/correlation/execute rules only if there is data (#94)

* adding DataTypes field

* if datatypes is empty do not execute the rule

* executing only if data exists for that rule

* fix frequence parsing (#96)

+++ correlation/sqldb/crud.go
@@ -35,7 +35,7 @@ func GetStatus() ([]map[string]interface{}, error) {
 	return dataSourceStatus, err
 }
 
-func UpdateStatistics(i, s, t string, c int) {
+func UpdateStatistics(i, s, t string, c int64) {
 	ping()
 
 	var err error
